fix: helper color function and rename color prop
