+++ src/components/Avatar/styled/container.js
@@ -1,12 +1,11 @@
 import styled from 'styled-components';
 import getTheme from '../../../styles/helpers/getTheme';
-import { validateColor } from '../../../styles/helpers/color';
+import { isValidColor } from '../../../styles/helpers/color';
 
 const StyledContainer = styled.span.attrs(props => {
     const theme = getTheme(props);
     const { getContrastText, brand } = theme.palette;
-    const isValidColor = validateColor(props.color);
-    const brandMainColor = isValidColor ? props.color : brand.main;
+    const brandMainColor = isValidColor(props.backgroundColor) ? props.backgroundColor : brand.main;
 
     return {
         brandMainColor,
