+++ src/styles/helpers/__test__/isValidColor.spec.js
@@ -1,10 +1,10 @@
-import { validateColor } from '../color';
+import { isValidColor } from '../color';
 
-describe('validateColor', () => {
+describe('isValidColor', () => {
     const colors = ['#01b6f5', 'rgb(1, 182, 245)', 'rgba(1, 182, 245, 0.5)'];
 
     it('should return true with a valid color', () => {
-        colors.forEach(color => expect(validateColor(color)).toBe(true));
+        colors.forEach(color => expect(isValidColor(color)).toBe(true));
     });
     it('should return false with an invalid color value', () => {
         [
@@ -17,6 +17,6 @@ describe('validateColor', () => {
             'rgb(1,2)',
             'rgb(20, 20, 20, 0.5)',
             'rgba(20, 20, 20)',
-        ].forEach(color => expect(validateColor(color)).toBe(false));
+        ].forEach(color => expect(isValidColor(color)).toBe(false));
     });
 });
