+++ src/styles/helpers/color/isValidColor.js
@@ -1,4 +1,4 @@
-export default function validateColor(color) {
+export default function isValidColor(color) {
     const element = document.createElement('a');
     element.style.color = color;
     return element.style.color !== '';
